server/requirements.txt
flask==3.1.0
pillow==11.1.0
google-generativeai==0.8.4
langchain-google-genai==2.0.9
langchain-core==0.3.34
faiss-cpu==1.10.0
sentence-transformers==3.4.1
# pickle5
numpy==2.2.2
